Power: Add support for Rittal 7955.310 PDUs
Like the ATEN this is controlled over SNMP.

I haven't written any tests for it.  I don't really believe in the value
of the existing ATEN tests so I don't want to duplicate them.  A better
approach might be to use [snmpsim] or similar to mock out the PDU at the
SNMP layer.

[snmpsim]: http://snmplabs.com/snmpsim/index.html